3 more misc tests
