+++ spec/api-web-contents-spec.js
@@ -98,137 +98,6 @@ describe('webContents module', () => {
     })
   })
 
-  describe('webframe messages in sandboxed contents', () => {
-    it('responds to executeJavaScript', async () => {
-      w.destroy()
-      w = new BrowserWindow({
-        show: false,
-        webPreferences: {
-          sandbox: true
-        }
-      })
-      await w.loadURL('about:blank')
-      const result = await w.webContents.executeJavaScript('37 + 5')
-      expect(result).to.equal(42)
-    })
-  })
-
-  describe('preload-error event', () => {
-    const generateSpecs = (description, sandbox) => {
-      describe(description, () => {
-        it('is triggered when unhandled exception is thrown', async () => {
-          const preload = path.join(fixtures, 'module', 'preload-error-exception.js')
-
-          w.destroy()
-          w = new BrowserWindow({
-            show: false,
-            webPreferences: {
-              sandbox,
-              preload
-            }
-          })
-
-          const promise = emittedOnce(w.webContents, 'preload-error')
-          w.loadURL('about:blank')
-
-          const [, preloadPath, error] = await promise
-          expect(preloadPath).to.equal(preload)
-          expect(error.message).to.equal('Hello World!')
-        })
-
-        it('is triggered on syntax errors', async () => {
-          const preload = path.join(fixtures, 'module', 'preload-error-syntax.js')
-
-          w.destroy()
-          w = new BrowserWindow({
-            show: false,
-            webPreferences: {
-              sandbox,
-              preload
-            }
-          })
-
-          const promise = emittedOnce(w.webContents, 'preload-error')
-          w.loadURL('about:blank')
-
-          const [, preloadPath, error] = await promise
-          expect(preloadPath).to.equal(preload)
-          expect(error.message).to.equal('foobar is not defined')
-        })
-
-        it('is triggered when preload script loading fails', async () => {
-          const preload = path.join(fixtures, 'module', 'preload-invalid.js')
-
-          w.destroy()
-          w = new BrowserWindow({
-            show: false,
-            webPreferences: {
-              sandbox,
-              preload
-            }
-          })
-
-          const promise = emittedOnce(w.webContents, 'preload-error')
-          w.loadURL('about:blank')
-
-          const [, preloadPath, error] = await promise
-          expect(preloadPath).to.equal(preload)
-          expect(error.message).to.contain('preload-invalid.js')
-        })
-      })
-    }
-
-    generateSpecs('without sandbox', false)
-    generateSpecs('with sandbox', true)
-  })
-
-  describe('takeHeapSnapshot()', () => {
-    it('works with sandboxed renderers', async () => {
-      w.destroy()
-      w = new BrowserWindow({
-        show: false,
-        webPreferences: {
-          sandbox: true
-        }
-      })
-
-      await w.loadURL('about:blank')
-
-      const filePath = path.join(remote.app.getPath('temp'), 'test.heapsnapshot')
-
-      const cleanup = () => {
-        try {
-          fs.unlinkSync(filePath)
-        } catch (e) {
-          // ignore error
-        }
-      }
-
-      try {
-        await w.webContents.takeHeapSnapshot(filePath)
-        const stats = fs.statSync(filePath)
-        expect(stats.size).not.to.be.equal(0)
-      } finally {
-        cleanup()
-      }
-    })
-
-    it('fails with invalid file path', async () => {
-      w.destroy()
-      w = new BrowserWindow({
-        show: false,
-        webPreferences: {
-          sandbox: true
-        }
-      })
-
-      await w.loadURL('about:blank')
-
-      const promise = w.webContents.takeHeapSnapshot('')
-      return expect(promise).to.be.eventually.rejectedWith(Error, 'takeHeapSnapshot failed')
-    })
-  })
-
   describe('setBackgroundThrottling()', () => {
     it('does not crash when allowing', (done) => {
       w.webContents.setBackgroundThrottling(true)
